Rename "settings" modules and variables to "tools"
