+++ examples/vanilla/related-genes.html
@@ -214,7 +214,7 @@
       annotationHeight: 5,
       onClickAnnot: onClickAnnot,
       onLoad: plotGeneFromUrl,
-      showSettings: true
+      showTools: true
     }
 
     let ideogram = Ideogram.initRelatedGenes(config)
