+++ src/js/tools/settings.js
